add retain_old_md config value and --retain-old-md cli option.
This adds the ability for createrepo to keep around the latest N old,
uniquely named, primary, filelist and otherdata xml and sqlite files
if they exist. This is handy for sites which have mirrors which are
not always in sync. It means if a user has an old repomd.xml they
can still get the xml or sqlite file referred to in that from
the normal mirrors