Exclude filename from .gz files
When you run createrepo, the original filenames,
including full path, of the files are stored
in the header of the .gz metadata

The path becomes part of the checksum of the gzip file
itself.  So, if you gunzip the file and re-gzip it,
you get a different checksum.

Based on the original observation of Dennis Gregorovic